refactor: move definitions to frontend
